Add shared scroller utilities for navigation and pause buttons
